First Commit
So far I have made an HTML file that accepts a battlescribe HTML list as an input. Added JS functionality, now it takes the inputted HTML list and changes the HTML on the page to display the list contents. This is the original output of battlescribe, next is to add JS that changes the layout of the html elements.